+++ packages/web-tracing/src/instrumentation.ts
@@ -7,7 +7,7 @@ import { BatchSpanProcessor } from '@opentelemetry/sdk-trace-base';
 import { WebTracerProvider } from '@opentelemetry/sdk-trace-web';
 import { SemanticResourceAttributes } from '@opentelemetry/semantic-conventions';
 
-import { BaseInstrumentation, VERSION } from '@grafana/faro-web-sdk';
+import { BaseInstrumentation, Transport, VERSION } from '@grafana/faro-web-sdk';
 
 import { FaroTraceExporter } from './faroTraceExporter';
 import { getDefaultOTELInstrumentations } from './getDefaultOTELInstrumentations';
@@ -75,6 +75,6 @@ export class TracingInstrumentation extends BaseInstrumentation {
   }
 
   private getIgnoreUrls(): Array<string | RegExp> {
-    return this.transports.transports.flatMap((transport) => transport.getIgnoreUrls());
+    return this.transports.transports.flatMap((transport: Transport) => transport.getIgnoreUrls());
   }
 }