+++ rollup.config.base.js
@@ -40,6 +40,12 @@ const modules = {
     globalName: 'GrafanaFaroInstrumentationPerformanceTimeline',
     externals: [],
   },
+  instrumentationFetch: {
+    name: '@grafana/faro-instrumentation-fetch',
+    bundleName: 'faro-instrumentation-fetch',
+    globalName: 'GrafanaFaroInstrumentationFetch',
+    externals: [],
+  },
 };
 
 exports.getRollupConfigBase = (moduleName) => {